services/api/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gridiron-sherlock-api"
version = "0.1.0"
description = "GridIron Sherlock API service skeleton"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.112,<0.116",
    "uvicorn[standard]>=0.30,<0.31",
    "pydantic>=2.7,<3.0",
    "pydantic-settings>=2.4,<3.0",
    "alembic>=1.13,<2.0",
    "sqlalchemy>=2.0,<3.0",
    "jsonschema>=4.22,<5.0",
    "email-validator>=2.2",
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3",
    "pytest-asyncio>=0.23",
    "mypy>=1.10",
    "httpx>=0.27",
    "urllib3>=2.0,<3.0",
]

[tool.setuptools.package-dir]
"services" = ".."

[tool.setuptools]
packages = [
    "services",
    "services.api",
    "services.api.app",
    "services.api.app.middleware",
    "services.api.app.routers",
    "services.api.app.db",
]
include-package-data = true
